// src/ext/annonars-api/src/lib/types.gen.ts
export type ClinvarAggregateGermlineReviewStatus =
  | 'no_classification_provided'
  | 'no_assertion_criteria_provided'
  | 'criteria_provided_single_submitter'
  | 'criteria_provided_multiple_submitters_no_conflicts'
  | 'criteria_provided_conflicting_classifications'
  | 'reviewed_by_expert_panel'
  | 'practice_guideline'
  | 'no_classifications_from_unflagged_records'
  | 'no_classification_for_the_single_variant'

/**
 * Enumeration describing aggregate oncogenicity review status value.
 */
export type ClinvarAggregateOncogenicityReviewStatus =
  | 'no_classification_provided'
  | 'no_assertion_criteria_provided'
  | 'criteria_provided_single_submitter'
  | 'criteria_provided_multiple_submitters_no_conflicts'
  | 'criteria_provided_conflicting_classifications'
  | 'reviewed_by_expert_panel'
  | 'practice_guideline'
  | 'no_classifications_from_unflagged_records'
  | 'no_classification_for_the_single_variant'

/**
 * Enumeration describing aggregate somatic clinical impact review status value.
 */
export type ClinvarAggregateSomaticClinicalImpactReviewStatus =
  | 'no_classification_provided'
  | 'no_assertion_criteria_provided'
  | 'criteria_provided_single_submitter'
  | 'criteria_provided_multiple_submitters'
  | 'reviewed_by_expert_panel'
  | 'practice_guideline'
  | 'no_classifications_from_unflagged_records'
  | 'no_classification_for_the_single_variant'